refactor(controller): isolate gateway epp resources
